requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
numpy
pandas
tf-keras
